Introduce `janus_messages` crate
Adds a new crate that contains Janus' Distributed Aggregation Protocol
message definitions and support for encoding and decoding them to/from
the wire format. By default, `janus_messages` does not pull in any
cryptography dependencies in order to make it easy to use.

Message definitions are moved from `janus_core::message` and
`janus_server::message` and are now all exported from `janus_messages`.

The DAP message decoding feature is moved from a subcommand on
`janus_server/src/bin/janus_cli.rs` into a new tool, `dap-decode`, built
from the `janus_messages` crate.

The `janus_messages` crate is published to crates.io on Janus releases,
and the CI build verifies that `janus_messages` builds without any
features enabled.

Part of #518
