+++ integration_tests/src/daphne.rs
@@ -2,7 +2,7 @@
 
 use daphne::DapGlobalConfig;
 use interop_binaries::test_util::{await_http_server, load_zstd_compressed_docker_image};
-use janus_core::message::{HpkeAeadId, HpkeConfig, HpkeKdfId, HpkeKemId, Role};
+use janus_messages::{HpkeAeadId, HpkeConfig, HpkeKdfId, HpkeKemId, Role};
 use janus_server::task::{Task, VdafInstance};
 use portpicker::pick_unused_port;
 use rand::{thread_rng, Rng};
