+++ interop_binaries/Cargo.toml
@@ -31,6 +31,7 @@ lazy_static = { version = "1", optional = true }
 janus_client = { path = "../janus_client" }
 janus_collector = { path = "../janus_collector" }
 janus_core = { path = "../janus_core" }
+janus_messages = { path = "../janus_messages" }
 janus_server = { path = "../janus_server" }
 opentelemetry = { version = "0.17", features = ["metrics"] }
 prio = { version = "0.8.2", features = ["multithreaded"] }
