+++ janus_collector/Cargo.toml
@@ -16,6 +16,7 @@ test-util = []
 backoff = { version = "0.4.0", features = ["tokio"] }
 derivative = "2.2.0"
 janus_core = { version = "0.1", path = "../janus_core" }
+janus_messages = { version = "0.1", path = "../janus_messages" }
 prio = "0.8.2"
 reqwest = { version = "0.11.4", default-features = false, features = ["rustls-tls"] }
 retry-after = "0.3.1"
