+++ janus_core/Cargo.toml
@@ -10,7 +10,6 @@ repository = "https://github.com/divviup/janus"
 rust-version = "1.63"
 
 [features]
-database = ["dep:bytes", "dep:postgres-protocol", "dep:postgres-types"]
 test-util = [
     "dep:assert_matches",
     "dep:lazy_static",
@@ -26,20 +25,15 @@ test-util = [
 ]
 
 [dependencies]
-anyhow = "1"
 backoff = { version = "0.4.0", features = ["tokio"] }
-base64 = "0.13.0"
-bytes = { version = "1.2.1", optional = true }
 chrono = "0.4"
 derivative = "2.2.0"
 futures = "0.3.24"
 hex = "0.4"
 hpke-dispatch = "0.3.0"
+janus_messages = { version = "0.1", path = "../janus_messages" }
 kube = { version = "0.65", optional = true, default-features = false, features = ["client", "rustls-tls"] }
 k8s-openapi = { version = "0.13.1", optional = true, features = ["v1_20"] }  # keep this version in sync with what is referenced by the indirect dependency via `kube`
-num_enum = "0.5.6"
-postgres-protocol = { version = "0.6.4", optional = true }
-postgres-types = { version = "0.2.4", optional = true }
 prio = { version = "0.8.2", features = ["multithreaded"] }
 rand = "0.8"
 reqwest = { version = "0.11.4", default-features = false, features = ["rustls-tls"] }
