+++ janus_core/src/hpke.rs
@@ -1,11 +1,11 @@
 //! Encryption and decryption of messages using HPKE (RFC 9180).
 
-use crate::message::{
+use derivative::Derivative;
+use hpke_dispatch::{HpkeError, Kem, Keypair};
+use janus_messages::{
     Extension, HpkeAeadId, HpkeCiphertext, HpkeConfig, HpkeConfigId, HpkeKdfId, HpkeKemId,
     HpkePublicKey, Interval, Nonce, Role, TaskId,
 };
-use derivative::Derivative;
-use hpke_dispatch::{HpkeError, Kem, Keypair};
 use prio::codec::{encode_u16_items, Encode};
 use std::{fmt::Debug, str::FromStr};
 
@@ -18,22 +18,20 @@ pub enum Error {
     InvalidConfiguration(&'static str),
 }
 
-impl TryFrom<&HpkeConfig> for hpke_dispatch::Config {
-    type Error = Error;
-
-    fn try_from(config: &HpkeConfig) -> Result<Self, Self::Error> {
-        Ok(Self {
-            aead: (config.aead_id() as u16)
-                .try_into()
-                .map_err(|_| Self::Error::InvalidConfiguration("did not recognize aead"))?,
-            kdf: (config.kdf_id() as u16)
-                .try_into()
-                .map_err(|_| Self::Error::InvalidConfiguration("did not recognize kdf"))?,
-            kem: (config.kem_id() as u16)
-                .try_into()
-                .map_err(|_| Self::Error::InvalidConfiguration("did not recognize kem"))?,
-        })
-    }
+fn hpke_dispatch_config_from_hpke_config(
+    config: &HpkeConfig,
+) -> Result<hpke_dispatch::Config, Error> {
+    Ok(hpke_dispatch::Config {
+        aead: (config.aead_id() as u16)
+            .try_into()
+            .map_err(|_| Error::InvalidConfiguration("did not recognize aead"))?,
+        kdf: (config.kdf_id() as u16)
+            .try_into()
+            .map_err(|_| Error::InvalidConfiguration("did not recognize kdf"))?,
+        kem: (config.kem_id() as u16)
+            .try_into()
+            .map_err(|_| Error::InvalidConfiguration("did not recognize kem"))?,
+    })
 }
 
 /// Construct the HPKE associated data for sealing or opening data enciphered for a report or report
@@ -140,7 +138,7 @@ pub fn seal(
     plaintext: &[u8],
     associated_data: &[u8],
 ) -> Result<HpkeCiphertext, Error> {
-    let output = hpke_dispatch::Config::try_from(recipient_config)?.base_mode_seal(
+    let output = hpke_dispatch_config_from_hpke_config(recipient_config)?.base_mode_seal(
         recipient_config.public_key().as_bytes(),
         &application_info.0,
         plaintext,
@@ -164,7 +162,7 @@ pub fn open(
     ciphertext: &HpkeCiphertext,
     associated_data: &[u8],
 ) -> Result<Vec<u8>, Error> {
-    hpke_dispatch::Config::try_from(recipient_config)?
+    hpke_dispatch_config_from_hpke_config(recipient_config)?
         .base_mode_open(
             &recipient_private_key.0,
             ciphertext.encapsulated_context(),
@@ -205,7 +203,7 @@ pub fn generate_hpke_config_and_private_key(
 #[cfg(feature = "test-util")]
 pub mod test_util {
     use super::{generate_hpke_config_and_private_key, HpkePrivateKey};
-    use crate::message::{HpkeAeadId, HpkeConfig, HpkeConfigId, HpkeKdfId, HpkeKemId};
+    use janus_messages::{HpkeAeadId, HpkeConfig, HpkeConfigId, HpkeKdfId, HpkeKemId};
     use rand::{thread_rng, Rng};
 
     pub fn generate_test_hpke_config_and_private_key() -> (HpkeConfig, HpkePrivateKey) {
@@ -221,14 +219,12 @@ pub mod test_util {
 #[cfg(test)]
 mod tests {
     use super::{test_util::generate_test_hpke_config_and_private_key, HpkeApplicationInfo, Label};
-    use crate::{
-        hpke::{open, seal, HpkePrivateKey},
-        message::{
-            HpkeAeadId, HpkeCiphertext, HpkeConfig, HpkeConfigId, HpkeKdfId, HpkeKemId,
-            HpkePublicKey, Role,
-        },
-    };
+    use crate::hpke::{open, seal, HpkePrivateKey};
     use hpke_dispatch::{Kem, Keypair};
+    use janus_messages::{
+        HpkeAeadId, HpkeCiphertext, HpkeConfig, HpkeConfigId, HpkeKdfId, HpkeKemId, HpkePublicKey,
+        Role,
+    };
     use serde::Deserialize;
     use std::collections::HashSet;
