+++ janus_messages/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "janus_messages"
+version = "0.1.12"
+edition = "2021"
+description = "Distributed Aggregation Protocol message definitions used in Janus, the server powering ISRG's Divvi Up."
+documentation = "https://docs.rs/janus_messages"
+homepage = "https://divviup.org"
+license = "MPL-2.0"
+repository = "https://github.com/divviup/janus"
+rust-version = "1.63"
+
+[features]
+dap-decode-bin = ["dep:structopt"]
+
+test-util = []
+
+[dependencies]
+anyhow = "1"
+base64 = "0.13.0"
+derivative = "2.2.0"
+hex = "0.4"
+num_enum = "0.5.6"
+prio = { version = "0.8.2", default-features = false } # important: do not enable prio/crypto-dependencies by default!
+rand = "0.8"
+serde = { version = "1.0.144", features = ["derive"] }
+structopt = { version = "0.3.26", optional = true }
+thiserror = "1.0"
+
+[dev-dependencies]
+assert_matches = "1"
+tracing-log = "0.1.3"
+tracing-subscriber = { version = "0.3", features = ["std", "env-filter", "fmt"] }
+
+[[bin]]
+name = "dap-decode"
+path = "src/bin/dap_decode.rs"
+required-features = ["dap-decode-bin"]
