+++ janus_messages/src/lib.rs
@@ -1,0 +1,2292 @@
+//! Messages defined by the [Distributed Aggregation Protocol][dap] with serialization and
+//! deserialization support.
+//!
+//! [dap]: https://datatracker.ietf.org/doc/draft-ietf-ppm-dap/
+
+use anyhow::anyhow;
+use base64::display::Base64Display;
+use derivative::Derivative;
+use num_enum::TryFromPrimitive;
+use prio::codec::{decode_u16_items, encode_u16_items, CodecError, Decode, Encode};
+use rand::{thread_rng, Rng};
+use serde::{Deserialize, Serialize};
+use std::{
+    fmt::{Debug, Display, Formatter},
+    io::{Cursor, Read},
+    str::FromStr,
+};
+
+/// Errors returned by functions and methods in this module
+#[derive(Debug, thiserror::Error)]
+pub enum Error {
+    /// An illegal arithmetic operation on a [`Time`] or [`Duration`].
+    #[error("{0}")]
+    IllegalTimeArithmetic(&'static str),
+}
+
+/// DAP protocol message representing a duration with a resolution of seconds.
+#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
+pub struct Duration(u64);
+
+impl Duration {
+    pub const ZERO: Duration = Duration::from_seconds(0);
+
+    /// Create a duration representing the provided number of seconds.
+    pub const fn from_seconds(seconds: u64) -> Self {
+        Self(seconds)
+    }
+
+    /// Get the number of seconds this duration represents.
+    pub fn as_seconds(self) -> u64 {
+        self.0
+    }
+}
+
+impl Encode for Duration {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        self.0.encode(bytes);
+    }
+}
+
+impl Decode for Duration {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        Ok(Self(u64::decode(bytes)?))
+    }
+}
+
+impl Display for Duration {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{} seconds", self.0)
+    }
+}
+
+/// DAP protocol message representing an instant in time with a resolution of seconds.
+#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, PartialOrd, Ord)]
+pub struct Time(u64);
+
+impl Time {
+    /// Get the number of seconds from January 1st, 1970, at 0:00:00 UTC to the instant represented
+    /// by this [`Time`] (i.e., the Unix timestamp for the instant it represents).
+    pub fn as_seconds_since_epoch(&self) -> u64 {
+        self.0
+    }
+
+    /// Construct a [`Time`] representing the instant that is a given number of seconds after
+    /// January 1st, 1970, at 0:00:00 UTC (i.e., the instant with the Unix timestamp of
+    /// `timestamp`).
+    pub const fn from_seconds_since_epoch(timestamp: u64) -> Self {
+        Self(timestamp)
+    }
+}
+
+impl Display for Time {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+impl Encode for Time {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        self.0.encode(bytes);
+    }
+}
+
+impl Decode for Time {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        Ok(Self(u64::decode(bytes)?))
+    }
+}
+
+/// DAP protocol message representing a half-open interval of time with a resolution of seconds;
+/// the start of the interval is included while the end of the interval is excluded.
+#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq)]
+pub struct Interval {
+    /// The start of the interval.
+    start: Time,
+    /// The length of the interval.
+    duration: Duration,
+}
+
+impl Interval {
+    /// Create a new [`Interval`] from the provided start and duration.
+    pub fn new(start: Time, duration: Duration) -> Result<Self, Error> {
+        start
+            .0
+            .checked_add(duration.0)
+            .ok_or(Error::IllegalTimeArithmetic("duration overflows time"))?;
+
+        Ok(Self { start, duration })
+    }
+
+    /// Returns a [`Time`] representing the included start of this interval.
+    pub fn start(&self) -> Time {
+        self.start
+    }
+
+    /// Get the duration of this interval.
+    pub fn duration(&self) -> Duration {
+        self.duration
+    }
+}
+
+impl Encode for Interval {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        self.start.encode(bytes);
+        self.duration.encode(bytes);
+    }
+}
+
+impl Decode for Interval {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let start = Time::decode(bytes)?;
+        let duration = Duration::decode(bytes)?;
+
+        Self::new(start, duration).map_err(|e| CodecError::Other(Box::new(e)))
+    }
+}
+
+impl Display for Interval {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "start: {} duration: {}", self.start, self.duration)
+    }
+}
+
+/// DAP protocol message representing a nonce uniquely identifying a client report.
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, PartialOrd, Ord)]
+pub struct Nonce {
+    /// The time at which the report was generated.
+    time: Time,
+    /// A randomly generated value.
+    rand: [u8; 16],
+}
+
+impl Nonce {
+    /// Construct a nonce with the given time and random parts.
+    pub fn new(time: Time, rand: [u8; 16]) -> Nonce {
+        Nonce { time, rand }
+    }
+
+    /// Get the time component of a nonce.
+    pub fn time(&self) -> Time {
+        self.time
+    }
+
+    /// Get the random component of a nonce.
+    pub fn rand(&self) -> [u8; 16] {
+        self.rand
+    }
+}
+
+impl Display for Nonce {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}.{}", self.time, hex::encode(self.rand))
+    }
+}
+
+impl Encode for Nonce {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        self.time.encode(bytes);
+        bytes.extend_from_slice(&self.rand);
+    }
+}
+
+impl Decode for Nonce {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let time = Time::decode(bytes)?;
+        let mut rand = [0; 16];
+        bytes.read_exact(&mut rand)?;
+
+        Ok(Self { time, rand })
+    }
+}
+
+/// Checksum over DAP report nonces, defined in §4.4.4.3.
+#[derive(Copy, Clone, Debug, Default, Hash, PartialEq, Eq)]
+pub struct NonceChecksum([u8; 32]);
+
+impl From<[u8; 32]> for NonceChecksum {
+    fn from(checksum: [u8; 32]) -> Self {
+        Self(checksum)
+    }
+}
+
+impl AsRef<[u8]> for NonceChecksum {
+    fn as_ref(&self) -> &[u8] {
+        &self.0
+    }
+}
+
+impl AsMut<[u8]> for NonceChecksum {
+    fn as_mut(&mut self) -> &mut [u8] {
+        &mut self.0
+    }
+}
+
+impl Display for NonceChecksum {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", hex::encode(self.0))
+    }
+}
+
+impl Encode for NonceChecksum {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        bytes.extend_from_slice(&self.0)
+    }
+}
+
+impl Decode for NonceChecksum {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let mut checksum = Self::default();
+        bytes.read_exact(&mut checksum.0)?;
+
+        Ok(checksum)
+    }
+}
+
+/// DAP protocol message representing the different roles that participants can adopt.
+#[derive(Copy, Clone, Debug, PartialEq, Eq, TryFromPrimitive, Serialize, Deserialize)]
+#[repr(u8)]
+pub enum Role {
+    Collector = 0,
+    Client = 1,
+    Leader = 2,
+    Helper = 3,
+}
+
+impl Role {
+    /// True if this [`Role`] is one of the aggregators.
+    pub fn is_aggregator(&self) -> bool {
+        matches!(self, Role::Leader | Role::Helper)
+    }
+
+    /// If this [`Role`] is one of the aggregators, returns the index at which
+    /// that aggregator's message or data can be found in various lists, or
+    /// `None` if the role is not an aggregator.
+    pub fn index(&self) -> Option<usize> {
+        match self {
+            // draft-gpew-priv-ppm §4.2: the leader's endpoint MUST be the first
+            Role::Leader => Some(0),
+            Role::Helper => Some(1),
+            _ => None,
+        }
+    }
+
+    pub fn as_str(&self) -> &'static str {
+        match self {
+            Self::Collector => "collector",
+            Self::Client => "client",
+            Self::Leader => "leader",
+            Self::Helper => "helper",
+        }
+    }
+}
+
+#[derive(Debug, thiserror::Error)]
+#[error("unknown role {0}")]
+pub struct RoleParseError(String);
+
+impl FromStr for Role {
+    type Err = RoleParseError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "collector" => Ok(Self::Collector),
+            "client" => Ok(Self::Client),
+            "leader" => Ok(Self::Leader),
+            "helper" => Ok(Self::Helper),
+            _ => Err(RoleParseError(s.to_owned())),
+        }
+    }
+}
+
+impl Encode for Role {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        (*self as u8).encode(bytes);
+    }
+}
+
+impl Decode for Role {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let val = u8::decode(bytes)?;
+        Self::try_from(val)
+            .map_err(|_| CodecError::Other(anyhow!("unexpected Role value {}", val).into()))
+    }
+}
+
+impl Display for Role {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.as_str())
+    }
+}
+
+/// DAP protocol message representing an identifier for an HPKE config.
+#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
+pub struct HpkeConfigId(u8);
+
+impl Display for HpkeConfigId {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+impl Encode for HpkeConfigId {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        self.0.encode(bytes);
+    }
+}
+
+impl Decode for HpkeConfigId {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        Ok(Self(u8::decode(bytes)?))
+    }
+}
+
+impl From<u8> for HpkeConfigId {
+    fn from(value: u8) -> HpkeConfigId {
+        HpkeConfigId(value)
+    }
+}
+
+impl From<HpkeConfigId> for u8 {
+    fn from(id: HpkeConfigId) -> u8 {
+        id.0
+    }
+}
+
+/// DAP protocol message representing an identifier for a DAP task.
+#[derive(Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord, Serialize, Deserialize)]
+pub struct TaskId([u8; Self::ENCODED_LEN]);
+
+impl Debug for TaskId {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "TaskId({})",
+            base64::display::Base64Display::with_config(&self.0, base64::URL_SAFE_NO_PAD)
+        )
+    }
+}
+
+impl Display for TaskId {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "{}",
+            base64::display::Base64Display::with_config(&self.0, base64::URL_SAFE_NO_PAD)
+        )
+    }
+}
+
+impl Encode for TaskId {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        bytes.extend_from_slice(&self.0)
+    }
+}
+
+impl Decode for TaskId {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let mut decoded = [0u8; Self::ENCODED_LEN];
+        bytes.read_exact(&mut decoded)?;
+        Ok(Self(decoded))
+    }
+}
+
+impl TaskId {
+    /// ENCODED_LEN is the length of a task ID in bytes when encoded.
+    pub const ENCODED_LEN: usize = 32;
+
+    /// Get a reference to the task ID as a byte slice
+    pub fn as_bytes(&self) -> &[u8] {
+        &self.0
+    }
+
+    /// Generate a random [`TaskId`]
+    pub fn random() -> Self {
+        let mut buf = [0u8; Self::ENCODED_LEN];
+        thread_rng().fill(&mut buf);
+        Self(buf)
+    }
+
+    /// Construct a [`TaskId`] from a byte array.
+    pub fn new(buf: [u8; Self::ENCODED_LEN]) -> Self {
+        Self(buf)
+    }
+}
+
+/// DAP protocol message representing an HPKE key encapsulation mechanism.
+#[derive(Clone, Copy, Debug, PartialEq, Eq, TryFromPrimitive, Serialize, Deserialize)]
+#[repr(u16)]
+pub enum HpkeKemId {
+    /// NIST P-256 keys and HKDF-SHA256.
+    P256HkdfSha256 = 0x0010,
+    /// X25519 keys and HKDF-SHA256.
+    X25519HkdfSha256 = 0x0020,
+}
+
+impl Encode for HpkeKemId {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        (*self as u16).encode(bytes);
+    }
+}
+
+impl Decode for HpkeKemId {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let val = u16::decode(bytes)?;
+        Self::try_from(val)
+            .map_err(|_| CodecError::Other(anyhow!("unexpected HpkeKemId value {}", val).into()))
+    }
+}
+
+/// DAP protocol message representing an HPKE key derivation function.
+#[derive(Clone, Copy, Debug, PartialEq, Eq, TryFromPrimitive, Serialize, Deserialize)]
+#[repr(u16)]
+pub enum HpkeKdfId {
+    /// HMAC Key Derivation Function SHA256.
+    HkdfSha256 = 0x0001,
+    /// HMAC Key Derivation Function SHA384.
+    HkdfSha384 = 0x0002,
+    /// HMAC Key Derivation Function SHA512.
+    HkdfSha512 = 0x0003,
+}
+
+impl Encode for HpkeKdfId {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        (*self as u16).encode(bytes);
+    }
+}
+
+impl Decode for HpkeKdfId {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let val = u16::decode(bytes)?;
+        Self::try_from(val)
+            .map_err(|_| CodecError::Other(anyhow!("unexpected HpkeKdfId value {}", val).into()))
+    }
+}
+
+/// DAP protocol message representing an HPKE AEAD.
+#[derive(Clone, Copy, Debug, PartialEq, Eq, TryFromPrimitive, Serialize, Deserialize)]
+#[repr(u16)]
+pub enum HpkeAeadId {
+    /// AES-128-GCM.
+    Aes128Gcm = 0x0001,
+    /// AES-256-GCM.
+    Aes256Gcm = 0x0002,
+    /// ChaCha20Poly1305.
+    ChaCha20Poly1305 = 0x0003,
+}
+
+impl Encode for HpkeAeadId {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        (*self as u16).encode(bytes);
+    }
+}
+
+impl Decode for HpkeAeadId {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let val = u16::decode(bytes)?;
+        Self::try_from(val)
+            .map_err(|_| CodecError::Other(anyhow!("unexpected HpkeAeadId value {}", val).into()))
+    }
+}
+
+/// DAP protocol message representing an arbitrary extension included in a client report.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct Extension {
+    extension_type: ExtensionType,
+    extension_data: Vec<u8>,
+}
+
+impl Extension {
+    /// Construct an extension from its type and payload.
+    pub fn new(extension_type: ExtensionType, extension_data: Vec<u8>) -> Extension {
+        Extension {
+            extension_type,
+            extension_data,
+        }
+    }
+}
+
+impl Encode for Extension {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        self.extension_type.encode(bytes);
+        encode_u16_items(bytes, &(), &self.extension_data);
+    }
+}
+
+impl Decode for Extension {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let extension_type = ExtensionType::decode(bytes)?;
+        let extension_data = decode_u16_items(&(), bytes)?;
+
+        Ok(Self {
+            extension_type,
+            extension_data,
+        })
+    }
+}
+
+/// DAP protocol message representing the type of an extension included in a client report.
+#[derive(Clone, Copy, Debug, Eq, PartialEq, TryFromPrimitive)]
+#[repr(u16)]
+pub enum ExtensionType {
+    Tbd = 0,
+}
+
+impl Encode for ExtensionType {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        (*self as u16).encode(bytes);
+    }
+}
+
+impl Decode for ExtensionType {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let val = u16::decode(bytes)?;
+        Self::try_from(val).map_err(|_| {
+            CodecError::Other(anyhow!("unexpected ExtensionType value {}", val).into())
+        })
+    }
+}
+
+/// DAP protocol message representing an HPKE ciphertext.
+#[derive(Clone, Derivative, Eq, PartialEq)]
+#[derivative(Debug)]
+pub struct HpkeCiphertext {
+    /// An identifier of the HPKE configuration used to seal the message.
+    config_id: HpkeConfigId,
+    /// An encasulated HPKE context.
+    #[derivative(Debug = "ignore")]
+    encapsulated_context: Vec<u8>,
+    /// An HPKE ciphertext.
+    #[derivative(Debug = "ignore")]
+    payload: Vec<u8>,
+}
+
+impl HpkeCiphertext {
+    /// Construct a HPKE ciphertext message from its components.
+    pub fn new(
+        config_id: HpkeConfigId,
+        encapsulated_context: Vec<u8>,
+        payload: Vec<u8>,
+    ) -> HpkeCiphertext {
+        HpkeCiphertext {
+            config_id,
+            encapsulated_context,
+            payload,
+        }
+    }
+
+    /// Get the configuration identifier associated with this ciphertext.
+    pub fn config_id(&self) -> HpkeConfigId {
+        self.config_id
+    }
+
+    /// Get the encapsulated key from this ciphertext message.
+    pub fn encapsulated_context(&self) -> &[u8] {
+        &self.encapsulated_context
+    }
+
+    /// Get the encrypted payload from this ciphertext message.
+    pub fn payload(&self) -> &[u8] {
+        &self.payload
+    }
+}
+
+impl Encode for HpkeCiphertext {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        self.config_id.encode(bytes);
+        encode_u16_items(bytes, &(), &self.encapsulated_context);
+        encode_u16_items(bytes, &(), &self.payload);
+    }
+}
+
+impl Decode for HpkeCiphertext {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let config_id = HpkeConfigId::decode(bytes)?;
+        let encapsulated_context = decode_u16_items(&(), bytes)?;
+        let payload = decode_u16_items(&(), bytes)?;
+
+        Ok(Self {
+            config_id,
+            encapsulated_context,
+            payload,
+        })
+    }
+}
+
+/// DAP protocol message representing an HPKE public key.
+// TODO(#230): refactor HpkePublicKey & HpkeConfig to simplify usage
+#[derive(Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct HpkePublicKey(pub(crate) Vec<u8>);
+
+impl HpkePublicKey {
+    /// Construct a `HpkePublicKey` from its byte array form.
+    pub fn new(buf: Vec<u8>) -> HpkePublicKey {
+        HpkePublicKey(buf)
+    }
+
+    /// Return the contents of this public key.
+    pub fn as_bytes(&self) -> &[u8] {
+        &self.0
+    }
+}
+
+impl Encode for HpkePublicKey {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        encode_u16_items(bytes, &(), &self.0);
+    }
+}
+
+impl Decode for HpkePublicKey {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let key = decode_u16_items(&(), bytes)?;
+        Ok(Self(key))
+    }
+}
+
+impl Debug for HpkePublicKey {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", hex::encode(&self.0))
+    }
+}
+
+/// DAP protocol message representing an HPKE config.
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+pub struct HpkeConfig {
+    id: HpkeConfigId,
+    kem_id: HpkeKemId,
+    kdf_id: HpkeKdfId,
+    aead_id: HpkeAeadId,
+    public_key: HpkePublicKey,
+}
+
+impl HpkeConfig {
+    /// The media type associated with this protocol message.
+    pub const MEDIA_TYPE: &'static str = "application/dap-hpke-config";
+
+    /// Construct a HPKE configuration message from its components.
+    pub fn new(
+        id: HpkeConfigId,
+        kem_id: HpkeKemId,
+        kdf_id: HpkeKdfId,
+        aead_id: HpkeAeadId,
+        public_key: HpkePublicKey,
+    ) -> HpkeConfig {
+        HpkeConfig {
+            id,
+            kem_id,
+            kdf_id,
+            aead_id,
+            public_key,
+        }
+    }
+
+    /// Returns the HPKE config ID associated with this HPKE configuration.
+    pub fn id(&self) -> HpkeConfigId {
+        self.id
+    }
+
+    /// Retrieve the key encapsulation mechanism algorithm identifier associated with this HPKE
+    /// configuration.
+    pub fn kem_id(&self) -> HpkeKemId {
+        self.kem_id
+    }
+
+    /// Retrieve the key derivation function algorithm identifier associated with this HPKE
+    /// configuration.
+    pub fn kdf_id(&self) -> HpkeKdfId {
+        self.kdf_id
+    }
+
+    /// Retrieve the AEAD algorithm identifier associated with this HPKE configuration.
+    pub fn aead_id(&self) -> HpkeAeadId {
+        self.aead_id
+    }
+
+    /// Retrieve the public key from this HPKE configuration.
+    pub fn public_key(&self) -> &HpkePublicKey {
+        &self.public_key
+    }
+}
+
+impl Encode for HpkeConfig {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        self.id.encode(bytes);
+        self.kem_id.encode(bytes);
+        self.kdf_id.encode(bytes);
+        self.aead_id.encode(bytes);
+        self.public_key.encode(bytes);
+    }
+}
+
+impl Decode for HpkeConfig {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let id = HpkeConfigId::decode(bytes)?;
+        let kem_id = HpkeKemId::decode(bytes)?;
+        let kdf_id = HpkeKdfId::decode(bytes)?;
+        let aead_id = HpkeAeadId::decode(bytes)?;
+        let public_key = HpkePublicKey::decode(bytes)?;
+
+        Ok(Self {
+            id,
+            kem_id,
+            kdf_id,
+            aead_id,
+            public_key,
+        })
+    }
+}
+
+/// DAP protocol message representing a client report.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct Report {
+    task_id: TaskId,
+    nonce: Nonce,
+    extensions: Vec<Extension>,
+    encrypted_input_shares: Vec<HpkeCiphertext>,
+}
+
+impl Report {
+    /// The media type associated with this protocol message.
+    pub const MEDIA_TYPE: &'static str = "application/dap-report";
+
+    /// Construct a report from its components.
+    pub fn new(
+        task_id: TaskId,
+        nonce: Nonce,
+        extensions: Vec<Extension>,
+        encrypted_input_shares: Vec<HpkeCiphertext>,
+    ) -> Report {
+        Report {
+            task_id,
+            nonce,
+            extensions,
+            encrypted_input_shares,
+        }
+    }
+
+    /// Retrieve the task identifier from this report.
+    pub fn task_id(&self) -> TaskId {
+        self.task_id
+    }
+
+    /// Get this report's nonce.
+    pub fn nonce(&self) -> Nonce {
+        self.nonce
+    }
+
+    /// Get this report's extensions.
+    pub fn extensions(&self) -> &[Extension] {
+        &self.extensions
+    }
+
+    /// Get this report's encrypted input shares.
+    pub fn encrypted_input_shares(&self) -> &[HpkeCiphertext] {
+        &self.encrypted_input_shares
+    }
+}
+
+impl Encode for Report {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        self.task_id.encode(bytes);
+        self.nonce.encode(bytes);
+        encode_u16_items(bytes, &(), &self.extensions);
+        encode_u16_items(bytes, &(), &self.encrypted_input_shares);
+    }
+}
+
+impl Decode for Report {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let task_id = TaskId::decode(bytes)?;
+        let timestamp = Nonce::decode(bytes)?;
+        let extensions = decode_u16_items(&(), bytes)?;
+        let encrypted_input_shares = decode_u16_items(&(), bytes)?;
+
+        Ok(Self {
+            task_id,
+            nonce: timestamp,
+            extensions,
+            encrypted_input_shares,
+        })
+    }
+}
+
+/// DAP protocol message representing a request from the collector to the leader to provide
+/// aggregate shares for a given batch interval.
+#[derive(Clone, Derivative, PartialEq, Eq)]
+#[derivative(Debug)]
+pub struct CollectReq {
+    pub task_id: TaskId,
+    pub batch_interval: Interval,
+    #[derivative(Debug = "ignore")]
+    pub agg_param: Vec<u8>,
+}
+
+impl CollectReq {
+    /// The media type associated with this protocol message.
+    pub const MEDIA_TYPE: &'static str = "application/dap-collect-req";
+}
+
+impl Encode for CollectReq {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        self.task_id.encode(bytes);
+        self.batch_interval.encode(bytes);
+        encode_u16_items(bytes, &(), &self.agg_param);
+    }
+}
+
+impl Decode for CollectReq {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let task_id = TaskId::decode(bytes)?;
+        let batch_interval = Interval::decode(bytes)?;
+        let agg_param = decode_u16_items(&(), bytes)?;
+
+        Ok(Self {
+            task_id,
+            batch_interval,
+            agg_param,
+        })
+    }
+}
+
+/// DAP protocol message representing a leader's response to the collector's request to provide
+/// aggregate shares for a given batch interval.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct CollectResp {
+    pub encrypted_agg_shares: Vec<HpkeCiphertext>,
+}
+
+impl CollectResp {
+    /// The media type associated with this protocol message.
+    pub const MEDIA_TYPE: &'static str = "application/dap-collect-resp";
+}
+
+impl Encode for CollectResp {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        encode_u16_items(bytes, &(), &self.encrypted_agg_shares);
+    }
+}
+
+impl Decode for CollectResp {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let encrypted_agg_shares = decode_u16_items(&(), bytes)?;
+
+        Ok(Self {
+            encrypted_agg_shares,
+        })
+    }
+}
+
+/// DAP protocol message representing one aggregator's share of a single client report.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct ReportShare {
+    pub nonce: Nonce,
+    pub extensions: Vec<Extension>,
+    pub encrypted_input_share: HpkeCiphertext,
+}
+
+impl Encode for ReportShare {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        self.nonce.encode(bytes);
+        encode_u16_items(bytes, &(), &self.extensions);
+        self.encrypted_input_share.encode(bytes);
+    }
+}
+
+impl Decode for ReportShare {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let nonce = Nonce::decode(bytes)?;
+        let extensions = decode_u16_items(&(), bytes)?;
+        let encrypted_input_share = HpkeCiphertext::decode(bytes)?;
+
+        Ok(Self {
+            nonce,
+            extensions,
+            encrypted_input_share,
+        })
+    }
+}
+
+/// DAP protocol message representing the result of a preparation step in a VDAF evaluation.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct PrepareStep {
+    pub nonce: Nonce,
+    pub result: PrepareStepResult,
+}
+
+impl Encode for PrepareStep {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        self.nonce.encode(bytes);
+        self.result.encode(bytes);
+    }
+}
+
+impl Decode for PrepareStep {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let nonce = Nonce::decode(bytes)?;
+        let result = PrepareStepResult::decode(bytes)?;
+
+        Ok(Self { nonce, result })
+    }
+}
+
+/// DAP protocol message representing result-type-specific data associated with a preparation step
+/// in a VDAF evaluation. Included in a PrepareStep message.
+#[derive(Clone, Derivative, PartialEq, Eq)]
+#[derivative(Debug)]
+pub enum PrepareStepResult {
+    Continued(#[derivative(Debug = "ignore")] Vec<u8>), // content is a serialized preparation message
+    Finished,
+    Failed(ReportShareError),
+}
+
+impl Encode for PrepareStepResult {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        // The encoding includes an implicit discriminator byte, called PrepareStepResult in the
+        // DAP spec.
+        match self {
+            Self::Continued(vdaf_msg) => {
+                0u8.encode(bytes);
+                encode_u16_items(bytes, &(), vdaf_msg);
+            }
+            Self::Finished => 1u8.encode(bytes),
+            Self::Failed(error) => {
+                2u8.encode(bytes);
+                error.encode(bytes);
+            }
+        }
+    }
+}
+
+impl Decode for PrepareStepResult {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let val = u8::decode(bytes)?;
+        Ok(match val {
+            0 => Self::Continued(decode_u16_items(&(), bytes)?),
+            1 => Self::Finished,
+            2 => Self::Failed(ReportShareError::decode(bytes)?),
+            _ => return Err(CodecError::UnexpectedValue),
+        })
+    }
+}
+
+/// DAP protocol message representing an error while preparing a report share for aggregation.
+#[derive(Clone, Copy, Debug, PartialEq, Eq, TryFromPrimitive)]
+#[cfg_attr(feature = "database", derive(ToSql, FromSql))]
+#[repr(u8)]
+pub enum ReportShareError {
+    BatchCollected = 0,
+    ReportReplayed = 1,
+    ReportDropped = 2,
+    HpkeUnknownConfigId = 3,
+    HpkeDecryptError = 4,
+    VdafPrepError = 5,
+}
+
+impl Encode for ReportShareError {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        (*self as u8).encode(bytes);
+    }
+}
+
+impl Decode for ReportShareError {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let val = u8::decode(bytes)?;
+        Self::try_from(val).map_err(|_| {
+            CodecError::Other(anyhow!("unexpected ReportShareError value {}", val).into())
+        })
+    }
+}
+
+/// DAP protocol message representing an identifier for an aggregation job.
+#[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub struct AggregationJobId([u8; Self::ENCODED_LEN]);
+
+impl AggregationJobId {
+    pub fn as_bytes(&self) -> &[u8] {
+        &self.0
+    }
+}
+
+impl Debug for AggregationJobId {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "AggregationJobId({})",
+            Base64Display::with_config(&self.0, base64::URL_SAFE_NO_PAD)
+        )
+    }
+}
+
+impl Display for AggregationJobId {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "{}",
+            Base64Display::with_config(&self.0, base64::URL_SAFE_NO_PAD)
+        )
+    }
+}
+
+impl Encode for AggregationJobId {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        bytes.extend_from_slice(&self.0)
+    }
+}
+
+impl Decode for AggregationJobId {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let mut decoded = [0u8; 32];
+        bytes.read_exact(&mut decoded)?;
+        Ok(Self(decoded))
+    }
+}
+
+impl AggregationJobId {
+    /// ENCODED_LEN is the length of an aggregation job ID in bytes when encoded.
+    const ENCODED_LEN: usize = 32;
+
+    /// Generate a random [`AggregationJobId`]
+    pub fn random() -> Self {
+        let mut buf = [0u8; Self::ENCODED_LEN];
+        thread_rng().fill(&mut buf);
+        Self(buf)
+    }
+}
+
+/// DAP protocol message representing an aggregation initialization request from leader to helper.
+#[derive(Clone, Derivative, PartialEq, Eq)]
+#[derivative(Debug)]
+pub struct AggregateInitializeReq {
+    pub task_id: TaskId,
+    pub job_id: AggregationJobId,
+    #[derivative(Debug = "ignore")]
+    pub agg_param: Vec<u8>,
+    pub report_shares: Vec<ReportShare>,
+}
+
+impl AggregateInitializeReq {
+    /// The media type associated with this protocol message.
+    pub const MEDIA_TYPE: &'static str = "application/dap-aggregate-initialize-req";
+}
+
+impl Encode for AggregateInitializeReq {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        self.task_id.encode(bytes);
+        self.job_id.encode(bytes);
+        encode_u16_items(bytes, &(), &self.agg_param);
+        encode_u16_items(bytes, &(), &self.report_shares);
+    }
+}
+
+impl Decode for AggregateInitializeReq {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let task_id = TaskId::decode(bytes)?;
+        let job_id = AggregationJobId::decode(bytes)?;
+        let agg_param = decode_u16_items(&(), bytes)?;
+        let report_shares = decode_u16_items(&(), bytes)?;
+        Ok(AggregateInitializeReq {
+            task_id,
+            job_id,
+            agg_param,
+            report_shares,
+        })
+    }
+}
+
+/// DAP protocol message representing an aggregation initialization response from helper to leader.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct AggregateInitializeResp {
+    pub prepare_steps: Vec<PrepareStep>,
+}
+
+impl AggregateInitializeResp {
+    /// The media type associated with this protocol message.
+    pub const MEDIA_TYPE: &'static str = "application/dap-aggregate-initialize-resp";
+}
+
+impl Encode for AggregateInitializeResp {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        encode_u16_items(bytes, &(), &self.prepare_steps);
+    }
+}
+
+impl Decode for AggregateInitializeResp {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let prepare_steps = decode_u16_items(&(), bytes)?;
+        Ok(Self { prepare_steps })
+    }
+}
+
+/// DAP protocol message representing an aggregation continuation request from leader to helper.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct AggregateContinueReq {
+    pub task_id: TaskId,
+    pub job_id: AggregationJobId,
+    pub prepare_steps: Vec<PrepareStep>,
+}
+
+impl AggregateContinueReq {
+    /// The media type associated with this protocol message.
+    pub const MEDIA_TYPE: &'static str = "application/dap-aggregate-continue-req";
+}
+
+impl Encode for AggregateContinueReq {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        self.task_id.encode(bytes);
+        self.job_id.encode(bytes);
+        encode_u16_items(bytes, &(), &self.prepare_steps);
+    }
+}
+
+impl Decode for AggregateContinueReq {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let task_id = TaskId::decode(bytes)?;
+        let job_id = AggregationJobId::decode(bytes)?;
+        let prepare_steps = decode_u16_items(&(), bytes)?;
+        Ok(Self {
+            task_id,
+            job_id,
+            prepare_steps,
+        })
+    }
+}
+
+/// DAP protocol message representing an aggregation continue response from helper to leader.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct AggregateContinueResp {
+    pub prepare_steps: Vec<PrepareStep>,
+}
+
+impl AggregateContinueResp {
+    /// The media type associated with this protocol message.
+    pub const MEDIA_TYPE: &'static str = "application/dap-aggregate-continue-resp";
+}
+
+impl Encode for AggregateContinueResp {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        encode_u16_items(bytes, &(), &self.prepare_steps);
+    }
+}
+
+impl Decode for AggregateContinueResp {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let prepare_steps = decode_u16_items(&(), bytes)?;
+        Ok(Self { prepare_steps })
+    }
+}
+
+/// DAP protocol message representing a request from the leader to a helper to provide an
+/// encrypted aggregate of its share of data for a given batch interval.
+#[derive(Clone, Derivative, PartialEq, Eq)]
+#[derivative(Debug)]
+pub struct AggregateShareReq {
+    pub task_id: TaskId,
+    pub batch_interval: Interval,
+    #[derivative(Debug = "ignore")]
+    pub aggregation_param: Vec<u8>,
+    pub report_count: u64,
+    pub checksum: NonceChecksum,
+}
+
+impl AggregateShareReq {
+    /// The media type associated with this protocol message.
+    pub const MEDIA_TYPE: &'static str = "application/dap-aggregate-share-req";
+}
+
+impl Encode for AggregateShareReq {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        self.task_id.encode(bytes);
+        self.batch_interval.encode(bytes);
+        encode_u16_items(bytes, &(), &self.aggregation_param);
+        self.report_count.encode(bytes);
+        self.checksum.encode(bytes);
+    }
+}
+
+impl Decode for AggregateShareReq {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let task_id = TaskId::decode(bytes)?;
+        let batch_interval = Interval::decode(bytes)?;
+        let agg_param = decode_u16_items(&(), bytes)?;
+        let report_count = u64::decode(bytes)?;
+        let checksum = NonceChecksum::decode(bytes)?;
+
+        Ok(Self {
+            task_id,
+            batch_interval,
+            aggregation_param: agg_param,
+            report_count,
+            checksum,
+        })
+    }
+}
+
+/// DAP protocol message representing a helper's response to the leader's request to provide an
+/// encrypted aggregate of its share of data for a given batch interval.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct AggregateShareResp {
+    pub encrypted_aggregate_share: HpkeCiphertext,
+}
+
+impl AggregateShareResp {
+    /// The media type associated with this protocol message.
+    pub const MEDIA_TYPE: &'static str = "application/dap-aggregate-share-resp";
+}
+
+impl Encode for AggregateShareResp {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        self.encrypted_aggregate_share.encode(bytes);
+    }
+}
+
+impl Decode for AggregateShareResp {
+    fn decode(bytes: &mut Cursor<&[u8]>) -> Result<Self, CodecError> {
+        let encrypted_aggregate_share = HpkeCiphertext::decode(bytes)?;
+
+        Ok(Self {
+            encrypted_aggregate_share,
+        })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use assert_matches::assert_matches;
+    use prio::codec::{CodecError, Decode, Encode};
+    use std::io::Cursor;
+
+    fn roundtrip_encoding<T>(vals_and_encodings: &[(T, &str)])
+    where
+        T: Encode + Decode + core::fmt::Debug + Eq,
+    {
+        for (val, hex_encoding) in vals_and_encodings {
+            let mut encoded_val = Vec::new();
+            val.encode(&mut encoded_val);
+            let encoding = hex::decode(hex_encoding).unwrap();
+            assert_eq!(encoding, encoded_val);
+            let decoded_val = T::decode(&mut Cursor::new(&encoded_val)).unwrap();
+            assert_eq!(val, &decoded_val);
+        }
+    }
+
+    #[test]
+    fn roundtrip_duration() {
+        roundtrip_encoding(&[
+            (Duration::from_seconds(u64::MIN), "0000000000000000"),
+            (Duration::from_seconds(12345), "0000000000003039"),
+            (Duration::from_seconds(u64::MAX), "FFFFFFFFFFFFFFFF"),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_time() {
+        roundtrip_encoding(&[
+            (Time::from_seconds_since_epoch(u64::MIN), "0000000000000000"),
+            (Time::from_seconds_since_epoch(12345), "0000000000003039"),
+            (Time::from_seconds_since_epoch(u64::MAX), "FFFFFFFFFFFFFFFF"),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_interval() {
+        Interval::new(
+            Time::from_seconds_since_epoch(1),
+            Duration::from_seconds(u64::MAX),
+        )
+        .unwrap_err();
+
+        let encoded = Interval {
+            start: Time::from_seconds_since_epoch(1),
+            duration: Duration::from_seconds(u64::MAX),
+        }
+        .get_encoded();
+        assert_eq!(
+            encoded,
+            hex::decode(concat!(
+                "0000000000000001", // start
+                "FFFFFFFFFFFFFFFF", // duration))
+            ))
+            .unwrap()
+        );
+
+        assert_matches!(Interval::get_decoded(&encoded), Err(CodecError::Other(_)));
+
+        roundtrip_encoding(&[
+            (
+                Interval {
+                    start: Time::from_seconds_since_epoch(u64::MIN),
+                    duration: Duration::from_seconds(u64::MAX),
+                },
+                concat!(
+                    "0000000000000000", // start
+                    "FFFFFFFFFFFFFFFF", // duration
+                ),
+            ),
+            (
+                Interval {
+                    start: Time::from_seconds_since_epoch(54321),
+                    duration: Duration::from_seconds(12345),
+                },
+                concat!(
+                    "000000000000D431", // start
+                    "0000000000003039", // duration
+                ),
+            ),
+            (
+                Interval {
+                    start: Time::from_seconds_since_epoch(u64::MAX),
+                    duration: Duration::from_seconds(u64::MIN),
+                },
+                concat!(
+                    "FFFFFFFFFFFFFFFF", // start
+                    "0000000000000000", // duration
+                ),
+            ),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_nonce() {
+        roundtrip_encoding(&[
+            (
+                Nonce::new(
+                    Time::from_seconds_since_epoch(12345),
+                    [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16],
+                ),
+                concat!(
+                    "0000000000003039",                 // time
+                    "0102030405060708090a0b0c0d0e0f10", // rand
+                ),
+            ),
+            (
+                Nonce::new(
+                    Time::from_seconds_since_epoch(54321),
+                    [16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1],
+                ),
+                concat!(
+                    "000000000000D431",                 // time
+                    "100f0e0d0c0b0a090807060504030201", // rand
+                ),
+            ),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_role() {
+        roundtrip_encoding(&[
+            (Role::Collector, "00"),
+            (Role::Client, "01"),
+            (Role::Leader, "02"),
+            (Role::Helper, "03"),
+        ]);
+    }
+
+    #[test]
+    fn roundtrip_hpke_config_id() {
+        roundtrip_encoding(&[
+            (HpkeConfigId(u8::MIN), "00"),
+            (HpkeConfigId(10), "0A"),
+            (HpkeConfigId(u8::MAX), "FF"),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_task_id() {
+        roundtrip_encoding(&[
+            (
+                TaskId::new([u8::MIN; 32]),
+                "0000000000000000000000000000000000000000000000000000000000000000",
+            ),
+            (
+                TaskId::new([
+                    0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
+                    22, 23, 24, 25, 26, 27, 28, 29, 30, 31,
+                ]),
+                "000102030405060708090A0B0C0D0E0F101112131415161718191A1B1C1D1E1F",
+            ),
+            (
+                TaskId::new([u8::MAX; 32]),
+                "FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF",
+            ),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_hpke_kem_id() {
+        roundtrip_encoding(&[
+            (HpkeKemId::P256HkdfSha256, "0010"),
+            (HpkeKemId::X25519HkdfSha256, "0020"),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_hpke_kdf_id() {
+        roundtrip_encoding(&[
+            (HpkeKdfId::HkdfSha256, "0001"),
+            (HpkeKdfId::HkdfSha384, "0002"),
+            (HpkeKdfId::HkdfSha512, "0003"),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_hpke_aead_id() {
+        roundtrip_encoding(&[
+            (HpkeAeadId::Aes128Gcm, "0001"),
+            (HpkeAeadId::Aes256Gcm, "0002"),
+            (HpkeAeadId::ChaCha20Poly1305, "0003"),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_extension() {
+        roundtrip_encoding(&[
+            (
+                Extension::new(ExtensionType::Tbd, Vec::new()),
+                concat!(
+                    "0000", // extension_type
+                    concat!(
+                        // extension_data
+                        "0000", // length
+                        "",     // opaque data
+                    ),
+                ),
+            ),
+            (
+                Extension::new(ExtensionType::Tbd, Vec::from("0123")),
+                concat!(
+                    "0000", // extension_type
+                    concat!(
+                        // extension_data
+                        "0004",     // length
+                        "30313233", // opaque data
+                    ),
+                ),
+            ),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_extension_type() {
+        roundtrip_encoding(&[(ExtensionType::Tbd, "0000")])
+    }
+
+    #[test]
+    fn roundtrip_hpke_ciphertext() {
+        roundtrip_encoding(&[
+            (
+                HpkeCiphertext::new(HpkeConfigId::from(10), Vec::from("0123"), Vec::from("4567")),
+                concat!(
+                    "0A", // config_id
+                    concat!(
+                        // encapsulated_context
+                        "0004",     // length
+                        "30313233", // opaque data
+                    ),
+                    concat!(
+                        // payload
+                        "0004",     // length
+                        "34353637", // opaque data
+                    ),
+                ),
+            ),
+            (
+                HpkeCiphertext::new(HpkeConfigId::from(12), Vec::from("01234"), Vec::from("567")),
+                concat!(
+                    "0C", // config_id
+                    concat!(
+                        // encapsulated_context
+                        "0005",       // length
+                        "3031323334", // opaque data
+                    ),
+                    concat!(
+                        // payload
+                        "0003",   // length
+                        "353637", // opaque data
+                    ),
+                ),
+            ),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_hpke_public_key() {
+        roundtrip_encoding(&[
+            (
+                HpkePublicKey::new(Vec::new()),
+                concat!(
+                    "0000", // length
+                    "",     // opaque data
+                ),
+            ),
+            (
+                HpkePublicKey::new(Vec::from("0123456789abcdef")),
+                concat!(
+                    "0010",                             // length
+                    "30313233343536373839616263646566"  // opaque data
+                ),
+            ),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_hpke_config() {
+        roundtrip_encoding(&[
+            (
+                HpkeConfig::new(
+                    HpkeConfigId::from(12),
+                    HpkeKemId::P256HkdfSha256,
+                    HpkeKdfId::HkdfSha512,
+                    HpkeAeadId::Aes256Gcm,
+                    HpkePublicKey::new(Vec::new()),
+                ),
+                concat!(
+                    "0C",   // id
+                    "0010", // kem_id
+                    "0003", // kdf_id
+                    "0002", // aead_id
+                    concat!(
+                        // public_key
+                        "0000", // length
+                        "",     // opaque data
+                    )
+                ),
+            ),
+            (
+                HpkeConfig::new(
+                    HpkeConfigId::from(23),
+                    HpkeKemId::X25519HkdfSha256,
+                    HpkeKdfId::HkdfSha256,
+                    HpkeAeadId::ChaCha20Poly1305,
+                    HpkePublicKey::new(Vec::from("0123456789abcdef")),
+                ),
+                concat!(
+                    "17",   // id
+                    "0020", // kem_id
+                    "0001", // kdf_id
+                    "0003", // aead_id
+                    concat!(
+                        // public_key
+                        "0010",                             // length
+                        "30313233343536373839616263646566", // opaque data
+                    )
+                ),
+            ),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_report() {
+        roundtrip_encoding(&[
+            (
+                Report::new(
+                    TaskId::new([u8::MIN; 32]),
+                    Nonce::new(
+                        Time::from_seconds_since_epoch(12345),
+                        [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16],
+                    ),
+                    vec![],
+                    vec![],
+                ),
+                concat!(
+                    "0000000000000000000000000000000000000000000000000000000000000000", // task_id
+                    concat!(
+                        // nonce
+                        "0000000000003039",                 // time
+                        "0102030405060708090a0b0c0d0e0f10", // rand
+                    ),
+                    concat!(
+                        // extensions
+                        "0000", // length
+                    ),
+                    concat!(
+                        // encrypted_input_shares
+                        "0000", // length
+                    )
+                ),
+            ),
+            (
+                Report::new(
+                    TaskId::new([u8::MAX; 32]),
+                    Nonce::new(
+                        Time::from_seconds_since_epoch(54321),
+                        [16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1],
+                    ),
+                    vec![Extension::new(ExtensionType::Tbd, Vec::from("0123"))],
+                    vec![
+                        HpkeCiphertext::new(
+                            HpkeConfigId::from(42),
+                            Vec::from("012345"),
+                            Vec::from("543210"),
+                        ),
+                        HpkeCiphertext::new(
+                            HpkeConfigId::from(13),
+                            Vec::from("abce"),
+                            Vec::from("abfd"),
+                        ),
+                    ],
+                ),
+                concat!(
+                    "FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF", // task_id
+                    concat!(
+                        "000000000000D431",                 // time
+                        "100f0e0d0c0b0a090807060504030201", // rand
+                    ),
+                    concat!(
+                        // extensions
+                        "0008", // length
+                        concat!(
+                            "0000", // extension_type
+                            concat!(
+                                // extension_data
+                                "0004",     // length
+                                "30313233", // opaque data
+                            ),
+                        )
+                    ),
+                    concat!(
+                        // encrypted_input_shares
+                        "001E", // length
+                        concat!(
+                            "2A", // config_id
+                            concat!(
+                                // encapsulated_context
+                                "0006",         // length
+                                "303132333435"  // opaque data
+                            ),
+                            concat!(
+                                // payload
+                                "0006",         // length
+                                "353433323130", // opaque data
+                            ),
+                        ),
+                        concat!(
+                            "0D", // config_id
+                            concat!(
+                                // encapsulated_context
+                                "0004",     // length
+                                "61626365", // opaque data
+                            ),
+                            concat!(
+                                // payload
+                                "0004",     // length
+                                "61626664", // opaque data
+                            ),
+                        ),
+                    ),
+                ),
+            ),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_collect_req() {
+        roundtrip_encoding(&[
+            (
+                CollectReq {
+                    task_id: TaskId::new([u8::MIN; 32]),
+                    batch_interval: Interval::new(
+                        Time::from_seconds_since_epoch(54321),
+                        Duration::from_seconds(12345),
+                    )
+                    .unwrap(),
+                    agg_param: Vec::new(),
+                },
+                concat!(
+                    "0000000000000000000000000000000000000000000000000000000000000000", // task_id,
+                    concat!(
+                        // batch_interval
+                        "000000000000D431", // start
+                        "0000000000003039", // duration
+                    ),
+                    concat!(
+                        // agg_param
+                        "0000", // length
+                        "",     // opaque data
+                    ),
+                ),
+            ),
+            (
+                CollectReq {
+                    task_id: TaskId::new([13u8; 32]),
+                    batch_interval: Interval::new(
+                        Time::from_seconds_since_epoch(48913),
+                        Duration::from_seconds(44721),
+                    )
+                    .unwrap(),
+                    agg_param: Vec::from("012345"),
+                },
+                concat!(
+                    "0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D0D", // task_id
+                    concat!(
+                        // batch_interval
+                        "000000000000BF11", // start
+                        "000000000000AEB1", // duration
+                    ),
+                    concat!(
+                        // agg_param
+                        "0006",         // length
+                        "303132333435", // opaque data
+                    ),
+                ),
+            ),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_collect_resp() {
+        roundtrip_encoding(&[
+            (
+                CollectResp {
+                    encrypted_agg_shares: Vec::new(),
+                },
+                concat!(concat!(
+                    // encrypted_agg_shares
+                    "0000", // length
+                )),
+            ),
+            (
+                CollectResp {
+                    encrypted_agg_shares: vec![
+                        HpkeCiphertext::new(
+                            HpkeConfigId::from(10),
+                            Vec::from("0123"),
+                            Vec::from("4567"),
+                        ),
+                        HpkeCiphertext::new(
+                            HpkeConfigId::from(12),
+                            Vec::from("01234"),
+                            Vec::from("567"),
+                        ),
+                    ],
+                },
+                concat!(concat!(
+                    // encrypted_agg_shares
+                    "001A", // length
+                    concat!(
+                        "0A", // config_id
+                        concat!(
+                            // encapsulated_context
+                            "0004",     // length
+                            "30313233", // opaque data
+                        ),
+                        concat!(
+                            // payload
+                            "0004",     // length
+                            "34353637", // opaque data
+                        ),
+                    ),
+                    concat!(
+                        "0C", // config_id
+                        concat!(
+                            // encapsulated_context
+                            "0005",       // length
+                            "3031323334", // opaque data
+                        ),
+                        concat!(
+                            // payload
+                            "0003",   // length
+                            "353637", // opaque data
+                        ),
+                    )
+                )),
+            ),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_prepare_step() {
+        roundtrip_encoding(&[
+            (
+                PrepareStep {
+                    nonce: Nonce::new(
+                        Time::from_seconds_since_epoch(54372),
+                        [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16],
+                    ),
+                    result: PrepareStepResult::Continued(Vec::from("012345")),
+                },
+                concat!(
+                    concat!(
+                        // nonce
+                        "000000000000D464",                 // time
+                        "0102030405060708090a0b0c0d0e0f10", // rand
+                    ),
+                    "00", // prepare_step_result
+                    concat!(
+                        // vdaf_msg
+                        "0006",         // length
+                        "303132333435", // opaque data
+                    ),
+                ),
+            ),
+            (
+                PrepareStep {
+                    nonce: Nonce::new(
+                        Time::from_seconds_since_epoch(12345),
+                        [16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1],
+                    ),
+                    result: PrepareStepResult::Finished,
+                },
+                concat!(
+                    concat!(
+                        // nonce
+                        "0000000000003039",                 // time
+                        "100f0e0d0c0b0a090807060504030201", // rand
+                    ),
+                    "01", // prepare_step_result
+                ),
+            ),
+            (
+                PrepareStep {
+                    nonce: Nonce::new(Time::from_seconds_since_epoch(345078), [255; 16]),
+                    result: PrepareStepResult::Failed(ReportShareError::VdafPrepError),
+                },
+                concat!(
+                    concat!(
+                        // nonce
+                        "00000000000543F6",                 // time
+                        "FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF", // rand
+                    ),
+                    "02", // prepare_step_result
+                    "05", // report_share_error
+                ),
+            ),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_report_share_error() {
+        roundtrip_encoding(&[
+            (ReportShareError::BatchCollected, "00"),
+            (ReportShareError::ReportReplayed, "01"),
+            (ReportShareError::ReportDropped, "02"),
+            (ReportShareError::HpkeUnknownConfigId, "03"),
+            (ReportShareError::HpkeDecryptError, "04"),
+            (ReportShareError::VdafPrepError, "05"),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_aggregation_job_id() {
+        roundtrip_encoding(&[
+            (
+                AggregationJobId([u8::MIN; 32]),
+                "0000000000000000000000000000000000000000000000000000000000000000",
+            ),
+            (
+                AggregationJobId([
+                    0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
+                    22, 23, 24, 25, 26, 27, 28, 29, 30, 31,
+                ]),
+                "000102030405060708090A0B0C0D0E0F101112131415161718191A1B1C1D1E1F",
+            ),
+            (
+                AggregationJobId([u8::MAX; 32]),
+                "FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF",
+            ),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_aggregate_initialize_req() {
+        roundtrip_encoding(&[(
+            AggregateInitializeReq {
+                task_id: TaskId::new([u8::MAX; 32]),
+                job_id: AggregationJobId([u8::MIN; 32]),
+                agg_param: Vec::from("012345"),
+                report_shares: vec![
+                    ReportShare {
+                        nonce: Nonce::new(
+                            Time::from_seconds_since_epoch(54321),
+                            [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16],
+                        ),
+                        extensions: vec![Extension::new(ExtensionType::Tbd, Vec::from("0123"))],
+                        encrypted_input_share: HpkeCiphertext::new(
+                            HpkeConfigId::from(42),
+                            Vec::from("012345"),
+                            Vec::from("543210"),
+                        ),
+                    },
+                    ReportShare {
+                        nonce: Nonce::new(
+                            Time::from_seconds_since_epoch(73542),
+                            [16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1],
+                        ),
+                        extensions: vec![Extension::new(ExtensionType::Tbd, Vec::from("3210"))],
+                        encrypted_input_share: HpkeCiphertext::new(
+                            HpkeConfigId::from(13),
+                            Vec::from("abce"),
+                            Vec::from("abfd"),
+                        ),
+                    },
+                ],
+            },
+            concat!(
+                "FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF", // task_id
+                "0000000000000000000000000000000000000000000000000000000000000000", // job_id
+                concat!(
+                    // agg_param
+                    "0006",         // length
+                    "303132333435", // opaque data
+                ),
+                concat!(
+                    // report_shares
+                    "0062", // length
+                    concat!(
+                        concat!(
+                            // nonce
+                            "000000000000D431",                 // time
+                            "0102030405060708090a0b0c0d0e0f10", // rand
+                        ),
+                        concat!(
+                            // extensions
+                            "0008", // length
+                            concat!(
+                                "0000", // extension_type
+                                concat!(
+                                    // extension_data
+                                    "0004",     // length
+                                    "30313233", // opaque data
+                                ),
+                            ),
+                        ),
+                        concat!(
+                            // encrypted_input_share
+                            "2A", // config_id
+                            concat!(
+                                // encapsulated_context
+                                "0006",         // length
+                                "303132333435", // opaque data
+                            ),
+                            concat!(
+                                // payload
+                                "0006",         // length
+                                "353433323130", // opaque data
+                            ),
+                        ),
+                    ),
+                    concat!(
+                        concat!(
+                            // nonce
+                            "0000000000011F46",                 // time
+                            "100f0e0d0c0b0a090807060504030201", // rand
+                        ),
+                        concat!(
+                            // extensions
+                            "0008", // length
+                            concat!(
+                                "0000", // extension_type
+                                concat!(
+                                    // extension_data
+                                    "0004",     // length
+                                    "33323130", // opaque data
+                                ),
+                            ),
+                        ),
+                        concat!(
+                            "0D", // config_id
+                            concat!(
+                                // encapsulated_context
+                                "0004",     // length
+                                "61626365", // opaque data
+                            ),
+                            concat!(
+                                // payload
+                                "0004",     // length
+                                "61626664", // opaque data
+                            ),
+                        ),
+                    ),
+                ),
+            ),
+        )])
+    }
+
+    #[test]
+    fn roundtrip_aggregate_initialize_resp() {
+        roundtrip_encoding(&[
+            (
+                AggregateInitializeResp {
+                    prepare_steps: vec![],
+                },
+                concat!(concat!(
+                    // prepare_steps
+                    "0000", // length
+                ),),
+            ),
+            (
+                AggregateInitializeResp {
+                    prepare_steps: vec![
+                        PrepareStep {
+                            nonce: Nonce::new(
+                                Time::from_seconds_since_epoch(54372),
+                                [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16],
+                            ),
+                            result: PrepareStepResult::Continued(Vec::from("012345")),
+                        },
+                        PrepareStep {
+                            nonce: Nonce::new(
+                                Time::from_seconds_since_epoch(12345),
+                                [16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1],
+                            ),
+                            result: PrepareStepResult::Finished,
+                        },
+                    ],
+                },
+                concat!(concat!(
+                    //prepare_steps
+                    "003A", // length
+                    concat!(
+                        concat!(
+                            // nonce
+                            "000000000000D464",                 // time
+                            "0102030405060708090a0b0c0d0e0f10", // rand
+                        ),
+                        "00", // prepare_step_result
+                        concat!(
+                            // payload
+                            "0006",         // length
+                            "303132333435", // opaque data
+                        ),
+                    ),
+                    concat!(
+                        concat!(
+                            // nonce
+                            "0000000000003039",                 // time
+                            "100f0e0d0c0b0a090807060504030201", // rand
+                        ),
+                        "01", // prepare_step_result
+                    ),
+                )),
+            ),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_aggregate_continue_req() {
+        roundtrip_encoding(&[(
+            AggregateContinueReq {
+                task_id: TaskId::new([u8::MIN; 32]),
+                job_id: AggregationJobId([u8::MAX; 32]),
+                prepare_steps: vec![
+                    PrepareStep {
+                        nonce: Nonce::new(
+                            Time::from_seconds_since_epoch(54372),
+                            [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16],
+                        ),
+                        result: PrepareStepResult::Continued(Vec::from("012345")),
+                    },
+                    PrepareStep {
+                        nonce: Nonce::new(
+                            Time::from_seconds_since_epoch(12345),
+                            [16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1],
+                        ),
+                        result: PrepareStepResult::Finished,
+                    },
+                ],
+            },
+            concat!(
+                "0000000000000000000000000000000000000000000000000000000000000000", // task_id
+                "FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF", // job_id
+                concat!(
+                    // prepare_steps
+                    "003A", // length
+                    concat!(
+                        concat!(
+                            // nonce
+                            "000000000000D464",                 // time
+                            "0102030405060708090a0b0c0d0e0f10", // rand
+                        ),
+                        "00", // prepare_step_result
+                        concat!(
+                            // payload
+                            "0006",         // length
+                            "303132333435", // opaque data
+                        ),
+                    ),
+                    concat!(
+                        concat!(
+                            // nonce
+                            "0000000000003039",                 // time
+                            "100f0e0d0c0b0a090807060504030201", // rand
+                        ),
+                        "01", // prepare_step_result
+                    )
+                ),
+            ),
+        )])
+    }
+
+    #[test]
+    fn roundtrip_aggregate_continue_resp() {
+        roundtrip_encoding(&[
+            (
+                AggregateContinueResp {
+                    prepare_steps: vec![],
+                },
+                concat!(concat!(
+                    // prepare_steps
+                    "0000", // length
+                ),),
+            ),
+            (
+                AggregateContinueResp {
+                    prepare_steps: vec![
+                        PrepareStep {
+                            nonce: Nonce::new(
+                                Time::from_seconds_since_epoch(54372),
+                                [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16],
+                            ),
+                            result: PrepareStepResult::Continued(Vec::from("012345")),
+                        },
+                        PrepareStep {
+                            nonce: Nonce::new(
+                                Time::from_seconds_since_epoch(12345),
+                                [16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1],
+                            ),
+                            result: PrepareStepResult::Finished,
+                        },
+                    ],
+                },
+                concat!(concat!(
+                    //prepare_steps
+                    "003A", // length
+                    concat!(
+                        concat!(
+                            // nonce
+                            "000000000000D464",                 // time
+                            "0102030405060708090a0b0c0d0e0f10", // rand
+                        ),
+                        "00", // prepare_step_result
+                        concat!(
+                            // payload
+                            "0006",         // length
+                            "303132333435", // opaque data
+                        ),
+                    ),
+                    concat!(
+                        concat!(
+                            // nonce
+                            "0000000000003039",                 // time
+                            "100f0e0d0c0b0a090807060504030201", // rand
+                        ),
+                        "01", // prepare_step_result
+                    ),
+                )),
+            ),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_aggregate_share_req() {
+        roundtrip_encoding(&[
+            (
+                AggregateShareReq {
+                    task_id: TaskId::new([u8::MIN; 32]),
+                    batch_interval: Interval::new(
+                        Time::from_seconds_since_epoch(54321),
+                        Duration::from_seconds(12345),
+                    )
+                    .unwrap(),
+                    aggregation_param: vec![],
+                    report_count: 439,
+                    checksum: NonceChecksum::get_decoded(&[u8::MIN; 32]).unwrap(),
+                },
+                concat!(
+                    "0000000000000000000000000000000000000000000000000000000000000000", // task_id
+                    concat!(
+                        // batch_interval
+                        "000000000000D431", // start
+                        "0000000000003039", // duration
+                    ),
+                    concat!(
+                        // agg_param
+                        "0000", // length
+                        "",     // opaque data
+                    ),
+                    "00000000000001B7", // report_count
+                    "0000000000000000000000000000000000000000000000000000000000000000", // checksum
+                ),
+            ),
+            (
+                AggregateShareReq {
+                    task_id: TaskId::new([12u8; 32]),
+                    batch_interval: Interval::new(
+                        Time::from_seconds_since_epoch(50821),
+                        Duration::from_seconds(84354),
+                    )
+                    .unwrap(),
+                    aggregation_param: Vec::from("012345"),
+                    report_count: 8725,
+                    checksum: NonceChecksum::get_decoded(&[u8::MAX; 32]).unwrap(),
+                },
+                concat!(
+                    "0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C0C", // task_id
+                    concat!(
+                        // batch_interval
+                        "000000000000C685", // start
+                        "0000000000014982", // duration
+                    ),
+                    concat!(
+                        // agg_param
+                        "0006",         // length
+                        "303132333435", // opaque data
+                    ),
+                    "0000000000002215", // report_count
+                    "FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF", // checksum
+                ),
+            ),
+        ])
+    }
+
+    #[test]
+    fn roundtrip_aggregate_share_resp() {
+        roundtrip_encoding(&[
+            (
+                AggregateShareResp {
+                    encrypted_aggregate_share: HpkeCiphertext::new(
+                        HpkeConfigId::from(10),
+                        Vec::from("0123"),
+                        Vec::from("4567"),
+                    ),
+                },
+                concat!(concat!(
+                    // encrypted_aggregate_share
+                    "0A", // config_id
+                    concat!(
+                        // encapsulated_context
+                        "0004",     // length
+                        "30313233", // opaque data
+                    ),
+                    concat!(
+                        // payload
+                        "0004",     // length
+                        "34353637", // opaque data
+                    ),
+                )),
+            ),
+            (
+                AggregateShareResp {
+                    encrypted_aggregate_share: HpkeCiphertext::new(
+                        HpkeConfigId::from(12),
+                        Vec::from("01234"),
+                        Vec::from("567"),
+                    ),
+                },
+                concat!(concat!(
+                    // encrypted_aggregate_share
+                    "0C", // config_id
+                    concat!(
+                        // encapsulated_context
+                        "0005",       // length
+                        "3031323334", // opaque data
+                    ),
+                    concat!(
+                        "0003",   // length
+                        "353637", // opaque data
+                    ),
+                )),
+            ),
+        ])
+    }
+}
