+++ janus_server/src/lib.rs
@@ -4,7 +4,7 @@ pub mod aggregator;
 pub mod binary_utils;
 pub mod config;
 pub mod datastore;
-pub mod message;
+pub mod messages;
 pub mod metrics;
 pub mod task;
 pub mod trace;
